review: bool/int8 science msg fields, docker comment in rover_env
Address Roozki's review nits on #179: drillmotorstatus and vacuumstatus
become bool, drillstepperdir becomes int8 (consumers compare to 0/1 so
no behaviour change), and the install/setup.bash guard in
rover_env_common.sh gets a comment explaining it exists for docker /
clean workspaces.

+++ src/rover_hmi_core/src/science/science_module.cpp
@@ -98,7 +98,7 @@ QWidget* ScienceModule::createWidget(QWidget* parent)
     for (auto [b, dir] : dirs) {
         b->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Fixed);
         QObject::connect(b, &QPushButton::clicked, [this, dir = dir]() {
-            state_.drillstepperdir = static_cast<int16_t>(dir);
+            state_.drillstepperdir = static_cast<int8_t>(dir);
             updateDrillBtns();
             publish();
         });
